datasets/preprocess/inpainting/QwenImage/requirements_qwen_image.txt
torch
torchvision
transformers==4.51.3


git+https://github.com/huggingface/diffusers

opencv-python
pillow==9.5.0
imageio>=2.35.1
numpy==1.26.0
scipy
pandas
matplotlib
omegaconf
beartype
fsspec
safetensors
huggingface_hub